// goalise/src/types/api/matchesPast.ts
export interface IMatchesPast {
  id: number;
  homeTeam: {
    id: number;
    name: string;
    logoUrl: string;
    captainId: number;
  };
  awayTeam: {
    id: number;
    name: string;
    logoUrl: string;
    captainId: number;
  };
  date: "2025-07-30T18:47:17.109Z";
  homeTeamScore: number;
  awayTeamScore: number;
}